Move VisitzBaseEndpoint into Endpoints namespace
