+++ visitz/VisitzApi/Endpoints/Attachments/GetAttachmentDetailsEndpoint.cs
@@ -2,7 +2,6 @@ using System.Net;
 using System.Text.Json;
 using VisitzApi.Json;
 using VisitzApi.Models.Attachments;
-using VisitzApi.Requests;
 
 namespace VisitzApi.Endpoints.Attachments;
