+++ visitz/VisitzApi/Endpoints/Attachments/PostAttachmentEndpoint.cs
@@ -2,7 +2,6 @@ using System.Text.Json;
 using VisitzApi.Extensions;
 using VisitzApi.Json;
 using VisitzApi.Models.Attachments;
-using VisitzApi.Requests;
 
 namespace VisitzApi.Endpoints.Attachments;
