+++ visitz/VisitzApi/Endpoints/GetCaseloadEndpoint.cs
@@ -1,7 +1,6 @@
 using System.Text.Json;
 using VisitzApi.Json;
 using VisitzApi.Models.Caseload;
-using VisitzApi.Requests;
 
 #nullable enable
