+++ visitz/VisitzApi/Endpoints/Notes/GetNotesEndpoint.cs
@@ -2,7 +2,6 @@ using System.Text.Json;
 using System.Text.Json.Nodes;
 using VisitzApi.Json;
 using VisitzApi.Models;
-using VisitzApi.Requests;
 
 namespace VisitzApi.Endpoints.Notes
 {
